backend/pyproject.toml
[tool.poetry]
name = "beeapi"
version = "2.0.0"
description = "IoT Beehive Monitoring API"
authors = ["Iuri Soares <your.email@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9,<3.14"
fastapi = "0.109.1"
uvicorn = {extras = ["standard"], version = "0.27.0"}
asyncpg = "0.29.0"
pydantic = {extras = ["email"], version = "^2.5.3"}
python-multipart = "0.0.20"
firebase-admin = "6.1.0"
python-dotenv = "1.0.0"
requests = "2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-asyncio = "^0.21.0"
httpx = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"